+++ client/src/layouts/AuthedLayout.tsx
@@ -3,12 +3,14 @@ import {
   Outlet,
   useNavigation,
   ScrollRestoration,
+  Params,
 } from "react-router-dom";
 import Footer from "../components/Footer";
 import Navbar from "../components/Navbar";
 import { useAuth } from "../context/AuthContext";
 import LoadingBar from "react-top-loading-bar";
 import React, { ReactElement, useEffect, useState } from "react";
+import axios from "axios";
 
 const AuthedLayout = ({ children }: { children?: ReactElement }) => {
   const { authedUser } = useAuth();
@@ -46,3 +48,16 @@ const AuthedLayout = ({ children }: { children?: ReactElement }) => {
 };
 
 export default AuthedLayout;
+
+export const authedLoader = async ({
+  request,
+  params,
+}: {
+  request: Request;
+  params: Params;
+}) => {
+  const userResponse = await axios.get(`/api/users/user`);
+  const userData = userResponse.data;
+
+  return { userData };
+};
